feat: implement useBrightness hook for managing screen brightness state and logic; refactor BrightnessControl component
